cPatterns Listing 2

+++ ing Logiciel/TP cPatterns/src/Produit.java
@@ -2,7 +2,10 @@ public class Produit {
     private String  nom;
     private double prixAchat;
 
-    public Produit(String nom, int prixAchat) {
+ //   public Produit() {
+ //  }
+
+    public Produit(String nom, double prixAchat) {
         this.nom = nom;
         this.prixAchat = prixAchat;
     }
